requirements.txt
Django==3.0.6
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
